refactor(vm): split vm_helpers.rs into cohesive modules (§7-8)
Split the 1695-line vm_helpers.rs (over the 500-line limit) into five
cohesive modules, all under 500 lines:

- vm_helpers.rs (258)             - backtrace string/value construction
- vm_helpers_lazy.rs (452)        - lazy-list force entry points + reconcile
- vm_helpers_lazy_pull.rs (373)   - force_lazy_list_vm_n_inner + force_lazy_pipe
- vm_helpers_lazy_scan.rs (229)   - pull_source_element + force_scan_lazy_list
- vm_helpers_junction.rs (395)    - junction/smartmatch threading + seq helpers

All methods are impl Interpreter, so the split is visibility-only. A few
cross-module private helpers are bumped to pub(super). The fn multiset is
byte-identical. No behavior change.
