+++ src/vm.rs
@@ -145,6 +145,10 @@ mod vm_dispatch_helpers;
 mod vm_env_helpers;
 pub(crate) mod vm_flipflop_ops;
 mod vm_helpers;
+mod vm_helpers_junction;
+mod vm_helpers_lazy;
+mod vm_helpers_lazy_pull;
+mod vm_helpers_lazy_scan;
 pub(crate) mod vm_hyper_func;
 mod vm_hyper_method_ops;
 pub(crate) mod vm_hyper_ops;
