+++ src/vm/vm_helpers.rs
@@ -255,1441 +255,4 @@ impl Interpreter {
         bt_attrs.insert("text".to_string(), Value::str(bt_str.to_string()));
         Value::make_instance(Symbol::intern("Backtrace"), bt_attrs)
     }
-
-    /// If the value is a `LazyIoLines`, force it into an eager array by reading
-    /// all remaining lines from the file handle. Otherwise return the value as-is.
-    pub(super) fn force_if_lazy_io_lines(&mut self, val: Value) -> Result<Value, RuntimeError> {
-        if let Value::LazyIoLines {
-            ref handle,
-            kv,
-            words,
-        } = val
-        {
-            let forced = loan_env!(self, force_lazy_io_lines(handle, words))?;
-            if kv {
-                // Apply .kv transformation on the forced array
-                let items = crate::runtime::utils::value_to_list(&forced);
-                let mut kv_items = Vec::with_capacity(items.len() * 2);
-                for (i, v) in items.iter().enumerate() {
-                    kv_items.push(Value::Int(i as i64));
-                    kv_items.push(v.clone());
-                }
-                Ok(Value::array(kv_items))
-            } else {
-                Ok(forced)
-            }
-        } else {
-            Ok(val)
-        }
-    }
-
-    pub(super) fn thread_right_first(
-        left: &crate::value::JunctionKind,
-        right: &crate::value::JunctionKind,
-    ) -> bool {
-        use crate::value::JunctionKind::{All, Any, None, One};
-        matches!(left, Any | One) && matches!(right, All | None)
-    }
-
-    pub(super) fn label_matches(error_label: &Option<String>, loop_label: &Option<String>) -> bool {
-        error_label.as_deref() == loop_label.as_deref() || error_label.is_none()
-    }
-
-    /// Check if a method on LazyList requires forcing the list first.
-    /// Methods that coerce a lazy `Seq` to another lazy view and so must
-    /// preserve the laziness of a map/grep pipeline (return it unchanged)
-    /// instead of forcing it. `.eager`/`.elems`/`.sort`/… are NOT here — those
-    /// genuinely need the whole list and correctly raise X::Cannot::Lazy.
-    pub(crate) fn lazy_pipe_preserving_coercion(method: &str) -> bool {
-        matches!(
-            method,
-            "Seq" | "List" | "list" | "Array" | "cache" | "values" | "lazy"
-        )
-    }
-
-    /// If the variable `name` holds a lazy `@`-array backed by a cache-bearing
-    /// spec (infinite sequence/closure/scan), reify its cached prefix into a
-    /// real Array and write it back, so a subsequent element mutation
-    /// (`@a[i] = v`, `:delete`) operates on a materialized backing. Front
-    /// mutations collapse the list to its prefix (no worse than the pre-L2
-    /// capped Array); reads keep it lazy. Reifies only cache-backed specs, so it
-    /// never runs user code or hangs. (L2)
-    pub(super) fn reify_lazy_array_slot(&mut self, name: &str) -> Result<(), RuntimeError> {
-        let lazy = match self.env().get(name) {
-            Some(Value::LazyList(ll))
-                if ll.in_array_context()
-                    && (ll.sequence_spec.is_some()
-                        || ll.closure_seq.is_some()
-                        || ll.scan_spec.is_some()) =>
-            {
-                Some(ll.clone())
-            }
-            _ => None,
-        };
-        if let Some(ll) = lazy {
-            // Reify a bounded prefix (matching the historical reify-to-cap of a
-            // capped `ArrayKind::Lazy` Array). With the L2b `[start]` seed the
-            // cache holds only one element, so `force_lazy_list_vm` (cache read)
-            // would lose the prefix on a front-mutation — extend to the cap.
-            const MAX_ARRAY_EXPAND: usize = 100_000;
-            let items = self.force_lazy_list_vm_n(&ll, MAX_ARRAY_EXPAND)?;
-            self.env_mut()
-                .insert(name.to_string(), Value::real_array(items));
-        }
-        Ok(())
-    }
-
-    pub(super) fn lazy_list_needs_forcing(method: &str) -> bool {
-        matches!(
-            method,
-            "list"
-                | "Array"
-                | "Numeric"
-                | "Int"
-                | "elems"
-                | "hyper"
-                | "race"
-                | "first"
-                | "grep"
-                | "map"
-                | "sort"
-                | "reverse"
-                | "join"
-                | "head"
-                | "tail"
-                | "min"
-                | "max"
-                | "minmax"
-                | "sum"
-                | "flat"
-                | "unique"
-                | "repeated"
-                | "squish"
-                | "classify"
-                | "categorize"
-                | "produce"
-                | "rotor"
-                | "batch"
-                | "reduce"
-                | "Supply"
-                | "combinations"
-                | "permutations"
-                | "values"
-                | "List"
-                | "Str"
-                | "Stringy"
-                | "gist"
-                | "raku"
-                | "perl"
-                | "Seq"
-                | "item"
-                | "cache"
-                | "pick"
-                | "roll"
-                | "keys"
-                | "kv"
-                | "pairs"
-                | "antipairs"
-        )
-    }
-
-    /// For `.head` on a gather-sourced `LazyList`, return how many leading
-    /// elements need to be produced (so an infinite gather is pulled lazily via
-    /// [`Self::force_lazy_list_vm_n`] instead of forced to completion).
-    /// Returns `None` for forms that need the whole list (e.g. `.head(*-3)`).
-    pub(super) fn gather_head_bound(method: &str, args: &[Value]) -> Option<usize> {
-        if method != "head" {
-            return None;
-        }
-        match args {
-            [] => Some(1),
-            [Value::Int(n)] => Some((*n).max(0) as usize),
-            [Value::Num(f)] => Some((*f as i64).max(0) as usize),
-            _ => None,
-        }
-    }
-
-    /// Force a LazyList by running its compiled bytecode in the Interpreter.
-    /// Falls back to interpreter if no compiled code is available.
-    pub(crate) fn force_lazy_list_vm(
-        &mut self,
-        list: &LazyList,
-    ) -> Result<Vec<Value>, RuntimeError> {
-        let caller_code = self.current_code;
-        let r = self.force_lazy_list_vm_inner(list);
-        self.reconcile_caller_after_lazy_force(caller_code);
-        r
-    }
-
-    fn force_lazy_list_vm_inner(&mut self, list: &LazyList) -> Result<Vec<Value>, RuntimeError> {
-        // Handle scan-based lazy lists: compute elements on demand
-        if list.scan_spec.is_some() {
-            return self.force_scan_lazy_list(list, 200_000);
-        }
-
-        // A lazy map/grep pipeline is rooted at an infinite source. It can still
-        // be forced when the callback runs `last` (which ends the sequence) —
-        // e.g. `(^Inf).grep({ last if $_ > 5; True }).eager`. So attempt a
-        // bounded force: if the pipe terminates within the cap (via `last` or a
-        // finite source), return it; otherwise it is genuinely infinite and we
-        // throw X::Cannot::Lazy, matching raku. Methods that know their own name
-        // (e.g. `.elems`/`.sort`) raise a more specific message at the dispatch
-        // site before reaching here.
-        if list.lazy_pipe.is_some() {
-            const EAGER_FORCE_CAP: usize = 1_000_000;
-            let forced = self.force_lazy_pipe(list, EAGER_FORCE_CAP)?;
-            let done = list
-                .lazy_pipe
-                .as_ref()
-                .map(|p| p.lock().unwrap().done)
-                .unwrap_or(true);
-            if done {
-                return Ok(forced);
-            }
-            return Err(RuntimeError::typed_msg(
-                "X::Cannot::Lazy",
-                "Cannot coerce an infinite lazy list to a strict list",
-            ));
-        }
-
-        // For sequence-spec lazy lists, a strict force cannot materialize the
-        // infinite tail, so return a bounded prefix (the historical 100k cap).
-        // With the L2b `[start]` seed the cache is O(1), so extend it to the cap
-        // here — every strict-force caller (front mutation reify, eager coerce,
-        // ...) expects the prefix, not the seed. Lazy *read* paths (index, head,
-        // first, map/grep pipes) use `force_lazy_list_vm_n` / pull and stay O(1).
-        if let Some(ref spec) = list.sequence_spec {
-            const MAX_ARRAY_EXPAND: usize = 100_000;
-            return Self::extend_sequence_cache(list, spec, MAX_ARRAY_EXPAND);
-        }
-
-        // Check cache first
-        if let Some(cached) = list.cache.lock().unwrap().clone() {
-            return Ok(cached);
-        }
-
-        // If no compiled code, fall back to interpreter
-        let (cc, fns) = match (&list.compiled_code, &list.compiled_fns) {
-            (Some(cc), Some(fns)) => (cc.clone(), fns.clone()),
-            _ => return self.force_lazy_list_bridge(list),
-        };
-
-        // Save current Interpreter state. Locals are kept coherent with env by
-        // write-through (`flush_local_to_env`), so no explicit flush is needed
-        // here; we restore locals directly on return.
-        crate::vm::vm_stats::record_clone_env();
-        let saved_env = self.clone_env();
-        let saved_locals = std::mem::take(&mut self.locals);
-        let saved_stack = std::mem::take(&mut self.stack);
-
-        // Set up the lazy list's environment as a scoped overlay's parent: the
-        // gather body reads its captured lexicals through to `list.env` and its
-        // own writes land in a fresh born-owned overlay (no fork of `list.env`).
-        // The merge below iterates the overlay (overlay-only) = the body's writes.
-        // See docs/vm-dual-store.md (Slice 6).
-        *self.env_mut() = crate::env::Env::scoped_child(list.env.flattened());
-
-        // Push gather items collector
-        let saved_gather_len = self.gather_items_len();
-        self.push_gather_items(Vec::new());
-        self.push_gather_take_limit(None);
-
-        // Initialize locals for the compiled code
-        self.locals = vec![Value::Nil; cc.locals.len()];
-        for (i, name) in cc.locals.iter().enumerate() {
-            if let Some(val) = self.env().get(name) {
-                self.locals[i] = val.clone();
-            }
-        }
-        self.stack = Vec::new();
-
-        // Run the compiled code using the lazy list's own compiled_fns.
-        // Outer scope subs are available via the env as Value::Sub.
-        let run_fns = fns.as_ref();
-
-        let mut ip = 0;
-        let mut run_result = Ok(());
-        while ip < cc.ops.len() {
-            match self.exec_one(&cc, &mut ip, run_fns) {
-                Ok(()) => {}
-                Err(e) if e.is_warn() => {
-                    if !self.warning_suppressed() {
-                        self.write_warn_to_stderr(&e.message);
-                    }
-                    if let Some(v) = e.return_value {
-                        self.stack.push(v);
-                    }
-                    ip += 1;
-                    continue;
-                }
-                Err(e) => {
-                    run_result = Err(e);
-                    break;
-                }
-            }
-            if self.is_halted() {
-                break;
-            }
-        }
-
-        // Collect gather items
-        let items = self.pop_gather_items().unwrap_or_default();
-        self.pop_gather_take_limit();
-
-        // Clean up extra gather items if needed
-        while self.gather_items_len() > saved_gather_len {
-            self.pop_gather_items();
-            self.pop_gather_take_limit();
-        }
-
-        // Sync locals back to env before reading the result environment.
-        // During Interpreter execution, variable assignments go to self.locals, not
-        // to the interpreter env. We must flush them so the merge logic below
-        // can see the changes made by the gather body.
-        for (i, name) in cc.locals.iter().enumerate() {
-            {
-                let __v = self.locals[i].clone();
-                self.env_mut().insert(name.clone(), __v);
-            }
-        }
-
-        // Restore the outer environment, selectively merging changes from
-        // the gather body. Only propagate variables that:
-        // 1. Existed in the outer scope, AND
-        // 2. Were actually modified during gather body execution
-        //    (i.e., their value changed from the gather body's initial env).
-        // This prevents nested gather closures from corrupting each other's
-        // captured variables (e.g., `$n` in nested grep-div calls), while
-        // still propagating genuine side effects (e.g., `$x += 1`).
-        let gather_result_env = self.env().clone();
-        let mut merged_env = saved_env.clone();
-        for (k, v) in gather_result_env.iter() {
-            if !saved_env.contains_key_sym(*k) {
-                continue;
-            }
-            if let Some(initial) = list.env.get_sym(*k) {
-                // Variable existed in both outer and gather env.
-                // Only propagate if the value actually changed during execution.
-                // Compare string representations as a proxy for value equality.
-                if v.to_string_value() != initial.to_string_value() {
-                    merged_env.insert_sym(*k, v.clone());
-                }
-            } else {
-                // Variable existed in outer scope but not in gather's initial env;
-                // always propagate changes.
-                merged_env.insert_sym(*k, v.clone());
-            }
-        }
-        *self.env_mut() = merged_env;
-        // Precise (env_dirty-independent) writeback: record the gather body's
-        // captured-outer writes so `apply_pending_rw_writeback` (called from
-        // `reconcile_caller_after_lazy_force`) drains them into the caller's local
-        // slots, exactly like a `map`/`grep` callback. Without this the gather
-        // case relied solely on the blanket `env_dirty` reconcile (the surface
-        // docs/captured-outer-cell-sharing.md is retiring).
-        self.record_eager_block_free_var_writeback(cc.as_ref(), &[]);
-
-        // Restore Interpreter state
-        self.locals = saved_locals;
-        self.stack = saved_stack;
-
-        // Check for errors
-        run_result?;
-
-        // Cache the result
-        *list.cache.lock().unwrap() = Some(items.clone());
-        Ok(items)
-    }
-
-    /// Force a gather-based LazyList to produce at least `needed` elements.
-    /// Uses coroutine-style suspend/resume: the gather body pauses at each
-    /// `take` once enough elements are available, and can be resumed later.
-    /// Side effects (e.g. `$count++`) are correctly scoped because we pause
-    /// mid-execution rather than re-running from scratch.
-    pub(super) fn force_lazy_list_vm_n(
-        &mut self,
-        list: &LazyList,
-        needed: usize,
-    ) -> Result<Vec<Value>, RuntimeError> {
-        let caller_code = self.current_code;
-        let r = self.force_lazy_list_vm_n_inner(list, needed);
-        self.reconcile_caller_after_lazy_force(caller_code);
-        r
-    }
-
-    /// Reconcile the caller frame's local slots after a lazy force, so a
-    /// captured-outer lexical mutated at reify time (e.g. `map({$c++})`,
-    /// `gather`) is visible in the caller's slots even when the blanket reverse
-    /// pull (`sync_locals_from_env`) is disabled (Slice F campaign). Two reify
-    /// shapes record their writes differently and both must be drained here —
-    /// the force machinery is the effective "call site" for the callbacks it
-    /// runs, but unlike a real call op it never drains them:
-    ///
-    /// - a `map`/`grep` callback runs via `call_compiled_closure`, which logs
-    ///   the changed caller free-vars into `pending_rw_writeback_sources`
-    ///   (drained by `apply_pending_rw_writeback`, as every call op does);
-    /// - a `gather` body merges its env changes and sets `env_dirty`, mirrored
-    ///   by `reconcile_locals_from_env_at_site` (the #3331 helper).
-    ///
-    /// Both are byte-identical to the work reverse-sync ON would do (the call-op
-    /// drain + the barrier pull), so ON behavior is unchanged.
-    ///
-    /// Also used by op handlers that run user code without a `code` parameter in
-    /// hand (`say`/`note`, which dispatch a `.gist`/`.Str` closure that can
-    /// mutate a captured-outer lexical): they capture `self.current_code` before
-    /// the dispatch and pass it here afterwards. `caller_code` is the address of
-    /// the caller frame's `CompiledCode`, captured before the inner `exec_one`
-    /// runs clobbered `current_code`.
-    pub(super) fn reconcile_caller_after_lazy_force(&mut self, caller_code: usize) {
-        // The force body's own `exec_one` runs reset `current_code` to the lazy
-        // body; restore it to the caller so a subsequent force in the same op
-        // handler reconciles the right frame.
-        self.current_code = caller_code;
-        if caller_code == 0 {
-            return;
-        }
-        if !self.pending_rw_writeback_sources.is_empty() {
-            // SAFETY: `caller_code` is the address of the `CompiledCode` of the
-            // bytecode frame that invoked this force. That frame is an ancestor
-            // on the call stack (the op handler driving the force) and is alive
-            // for the whole synchronous duration of the force, so the pointer is
-            // valid here.
-            let code = unsafe { &*(caller_code as *const CompiledCode) };
-            self.apply_pending_rw_writeback(code);
-        }
-    }
-
-    /// Reconcile after an *internal redispatch* (a user coercion/render method run
-    /// mid-opcode without a surrounding CallMethod op: `+$obj`/`~$obj`/`if $obj`,
-    /// string interpolation, `put`/`print`, …). Like
-    /// [`Self::reconcile_caller_after_lazy_force`] it drains the captured-outer
-    /// writeback into the caller frame's local slots, BUT it **retains** any
-    /// `pending_rw_writeback_sources` entry whose name is not a slot of
-    /// `caller_code` instead of dropping it.
-    ///
-    /// The retain matters because an internal redispatch can fire *inside another
-    /// method body that has not yet returned* — e.g. a `submethod BUILD`'s
-    /// `$gather ~= "($a)"` interpolation runs while a *sibling* `BUILD`'s
-    /// captured-outer write (`$parent-counter++`) is still sitting in
-    /// `pending_rw_writeback_sources`, queued for the outer `.new` call site to
-    /// drain. The drop-on-miss `apply_pending_rw_writeback` would consume and
-    /// discard that sibling write here (its slot lives in the outer frame, not the
-    /// BUILD frame), so the outer `.new` drain finds nothing and the caller's slot
-    /// stays stale. Retaining the miss leaves it for the frame that actually owns
-    /// the slot.
-    pub(super) fn reconcile_caller_after_internal_dispatch(&mut self, caller_code: usize) {
-        self.current_code = caller_code;
-        if caller_code == 0 {
-            return;
-        }
-        // SAFETY: see `reconcile_caller_after_lazy_force` — `caller_code` is the
-        // live ancestor frame's `CompiledCode` address.
-        let code = unsafe { &*(caller_code as *const CompiledCode) };
-        if !self.pending_rw_writeback_sources.is_empty() {
-            let sources = std::mem::take(&mut self.pending_rw_writeback_sources);
-            let mut retained = Vec::new();
-            for source in sources {
-                if let Some(slot) = self.find_local_slot(code, &source) {
-                    if !matches!(self.locals[slot], Value::HashEntryRef { .. })
-                        && let Some(val) = self.env().get(&source).cloned()
-                    {
-                        self.locals[slot] = val;
-                    }
-                    // matched (slot in this frame) → applied, do not retain
-                } else {
-                    retained.push(source);
-                }
-            }
-            self.pending_rw_writeback_sources = retained;
-        }
-        self.apply_pending_caller_var_writeback(code);
-    }
-
-    fn force_lazy_list_vm_n_inner(
-        &mut self,
-        list: &LazyList,
-        needed: usize,
-    ) -> Result<Vec<Value>, RuntimeError> {
-        // Check cache first
-        {
-            let cache = list.cache.lock().unwrap();
-            if let Some(cached) = cache.as_ref()
-                && cached.len() >= needed
-            {
-                return Ok(cached[..needed].to_vec());
-            }
-        }
-
-        // For sequence-spec lazy lists, generate more elements on demand
-        if let Some(ref spec) = list.sequence_spec {
-            return Self::extend_sequence_cache(list, spec, needed);
-        }
-
-        // For infinite closure-based sequences (`1, 1, * + * ... *`), re-invoke
-        // the generator closure over the growing history to produce elements on
-        // demand, so an unbounded sequence stays lazy instead of truncating to
-        // its eager prefix.
-        if list.closure_seq.is_some() {
-            return self.extend_closure_sequence(list, needed);
-        }
-
-        // Lazy map/grep pipeline: pull from the source and apply the stage on
-        // demand (one source element at a time), so an infinite source stays
-        // lazy instead of materializing.
-        if list.lazy_pipe.is_some() {
-            return self.force_lazy_pipe(list, needed);
-        }
-
-        // Check if coroutine is finished (body completed, all elements produced)
-        if let Some(ref coro_mutex) = list.coroutine {
-            let coro = coro_mutex.lock().unwrap();
-            if coro.finished {
-                // Body is done; return whatever we have cached
-                let cache = list.cache.lock().unwrap();
-                return Ok(cache.as_ref().cloned().unwrap_or_default());
-            }
-        }
-
-        // Need compiled code
-        let (cc, fns) = match (&list.compiled_code, &list.compiled_fns) {
-            (Some(cc), Some(fns)) => (cc.clone(), fns.clone()),
-            _ => {
-                // Fall back to interpreter prefix bridge
-                return self.force_lazy_list_prefix_bridge(list, needed);
-            }
-        };
-
-        // Save current Interpreter state
-        crate::vm::vm_stats::record_clone_env();
-        let saved_env = self.clone_env();
-        let saved_locals = std::mem::take(&mut self.locals);
-        let saved_stack = std::mem::take(&mut self.stack);
-
-        // Determine starting IP and locals from coroutine state or fresh start
-        let mut ip;
-        let has_prior_state;
-
-        if let Some(ref coro_mutex) = list.coroutine {
-            let coro = coro_mutex.lock().unwrap();
-            if !coro.finished && coro.ip > 0 {
-                // Resume from saved state
-                ip = coro.ip;
-                self.locals = coro.locals.clone();
-                self.stack = coro.stack.clone();
-                *self.env_mut() = coro.env.clone();
-                self.gather_for_loop_resume = coro.for_loop_resume.clone();
-                has_prior_state = true;
-            } else {
-                // Fresh start (scoped overlay over the gather's captured env; see
-                // docs/vm-dual-store.md Slice 6). On suspend the scoped env is
-                // saved into the coroutine state and restored on resume (clone
-                // preserves overlay+parent+tombstones), so the body's writes
-                // accumulate across takes without forking `list.env`.
-                ip = 0;
-                *self.env_mut() = crate::env::Env::scoped_child(list.env.flattened());
-                self.locals = vec![Value::Nil; cc.locals.len()];
-                for (i, name) in cc.locals.iter().enumerate() {
-                    if let Some(val) = self.env().get(name) {
-                        self.locals[i] = val.clone();
-                    }
-                }
-                self.stack = Vec::new();
-                has_prior_state = false;
-            }
-        } else {
-            // Fresh start (no coroutine slot yet)
-            ip = 0;
-            *self.env_mut() = crate::env::Env::scoped_child(list.env.flattened());
-            self.locals = vec![Value::Nil; cc.locals.len()];
-            for (i, name) in cc.locals.iter().enumerate() {
-                if let Some(val) = self.env().get(name) {
-                    self.locals[i] = val.clone();
-                }
-            }
-            self.stack = Vec::new();
-            has_prior_state = false;
-        }
-
-        // Push gather items collector with the take limit
-        let saved_gather_len = self.gather_items_len();
-
-        // If resuming, restore already-cached items into the gather collector
-        // so that the take_value limit check accounts for them.
-        let _already_cached = if has_prior_state {
-            let cache = list.cache.lock().unwrap();
-            let items = cache.as_ref().cloned().unwrap_or_default();
-            let len = items.len();
-            self.push_gather_items(items);
-            len
-        } else {
-            self.push_gather_items(Vec::new());
-            0
-        };
-        self.push_gather_take_limit(Some(needed));
-
-        // Run the compiled code
-        let run_fns = fns.as_ref();
-        let mut run_result = Ok(());
-        let mut body_finished = false;
-
-        while ip < cc.ops.len() {
-            match self.exec_one(&cc, &mut ip, run_fns) {
-                Ok(()) => {}
-                Err(e) if e.is_warn() => {
-                    if !self.warning_suppressed() {
-                        self.write_warn_to_stderr(&e.message);
-                    }
-                    if let Some(v) = e.return_value {
-                        self.stack.push(v);
-                    }
-                    ip += 1;
-                    continue;
-                }
-                Err(e)
-                    if e.message == crate::runtime::Interpreter::LAZY_GATHER_TAKE_LIMIT_SIGNAL =>
-                {
-                    // Take limit reached — the gather body yielded.
-                    if self.gather_for_loop_resume.is_some() {
-                        // From inside a compound ForLoop op — keep ip here.
-                    } else {
-                        // ip points to the Take instruction; advance past it.
-                        ip += 1;
-                    }
-                    break;
-                }
-                Err(e) => {
-                    run_result = Err(e);
-                    break;
-                }
-            }
-            if self.is_halted() {
-                break;
-            }
-        }
-
-        // If we exited the loop normally (ip >= cc.ops.len()), body is finished
-        if run_result.is_ok() && ip >= cc.ops.len() {
-            body_finished = true;
-        }
-
-        // Collect gather items
-        let items = self.pop_gather_items().unwrap_or_default();
-        self.pop_gather_take_limit();
-
-        while self.gather_items_len() > saved_gather_len {
-            self.pop_gather_items();
-            self.pop_gather_take_limit();
-        }
-
-        // Sync locals back to env
-        for (i, name) in cc.locals.iter().enumerate() {
-            {
-                let __v = self.locals[i].clone();
-                self.env_mut().insert(name.clone(), __v);
-            }
-        }
-
-        // Save coroutine state before restoring outer env
-        if !body_finished && run_result.is_ok() {
-            let for_loop_resume = self.gather_for_loop_resume.take();
-            let coro_state = GatherCoroutineState {
-                ip,
-                locals: self.locals.clone(),
-                stack: self.stack.clone(),
-                env: self.env().clone(),
-                finished: false,
-                for_loop_resume,
-            };
-            if let Some(ref coro_mutex) = list.coroutine {
-                *coro_mutex.lock().unwrap() = coro_state;
-            }
-            // Note: if list.coroutine is None, we can't save state (shouldn't happen
-            // for gather-based lists since we set it up in MakeGather)
-        } else if let Some(ref coro_mutex) = list.coroutine {
-            let mut coro = coro_mutex.lock().unwrap();
-            coro.finished = true;
-        }
-
-        // Merge env changes back to outer scope
-        let gather_result_env = self.env().clone();
-        let initial_env = &list.env;
-        let mut merged_env = saved_env.clone();
-        for (k, v) in gather_result_env.iter() {
-            if !saved_env.contains_key_sym(*k) {
-                continue;
-            }
-            if let Some(initial) = initial_env.get_sym(*k) {
-                if v.to_string_value() != initial.to_string_value() {
-                    merged_env.insert_sym(*k, v.clone());
-                }
-            } else {
-                merged_env.insert_sym(*k, v.clone());
-            }
-        }
-        *self.env_mut() = merged_env;
-        // Precise (env_dirty-independent) writeback of the gather body's
-        // captured-outer writes — see the matching comment in
-        // `force_lazy_list_vm_inner`.
-        self.record_eager_block_free_var_writeback(cc.as_ref(), &[]);
-
-        // Restore Interpreter state
-        self.locals = saved_locals;
-        self.stack = saved_stack;
-
-        run_result?;
-
-        // Update cache
-        *list.cache.lock().unwrap() = Some(items.clone());
-
-        let result = if items.len() > needed {
-            items[..needed].to_vec()
-        } else {
-            items
-        };
-        Ok(result)
-    }
-
-    /// Produce at least `needed` output elements of a lazy `map`/`grep` pipeline
-    /// (a `LazyList` carrying a [`crate::value::MapGrepSpec`]).
-    ///
-    /// Elements are produced by pulling from the stage's source one at a time
-    /// and applying the `map`/`grep` callback, accumulating outputs in the
-    /// list's cache. A `grep` stage filters (0 or 1 output per source element);
-    /// a `map` stage transforms (a `Slip` result contributes multiple). The
-    /// source itself may be another lazy pipeline, so chains nest.
-    pub(crate) fn force_lazy_pipe(
-        &mut self,
-        list: &LazyList,
-        needed: usize,
-    ) -> Result<Vec<Value>, RuntimeError> {
-        loop {
-            // Fast path: enough already cached, or the source is exhausted.
-            {
-                let cache = list.cache.lock().unwrap();
-                let done = list
-                    .lazy_pipe
-                    .as_ref()
-                    .map(|p| p.lock().unwrap().done)
-                    .unwrap_or(true);
-                if let Some(c) = cache.as_ref()
-                    && (c.len() >= needed || done)
-                {
-                    let n = needed.min(c.len());
-                    return Ok(c[..n].to_vec());
-                }
-            }
-
-            // Snapshot the stage so no pipe lock is held across the pull/apply
-            // (which may recursively pull from a nested pipe source).
-            let (source, func, is_grep, source_idx, index_transform) = {
-                let spec = list.lazy_pipe.as_ref().unwrap().lock().unwrap();
-                (
-                    spec.source.clone(),
-                    spec.func.clone(),
-                    spec.is_grep,
-                    spec.source_idx,
-                    spec.index_transform,
-                )
-            };
-
-            match self.pull_source_element(&source, source_idx)? {
-                None => {
-                    // Source exhausted: mark done and return what we have.
-                    let mut spec = list.lazy_pipe.as_ref().unwrap().lock().unwrap();
-                    spec.done = true;
-                    drop(spec);
-                    let cache = list.cache.lock().unwrap();
-                    let c = cache.as_ref().cloned().unwrap_or_default();
-                    let n = needed.min(c.len());
-                    return Ok(c[..n].to_vec());
-                }
-                Some(elem) if index_transform.is_some() => {
-                    // `.pairs`/`.antipairs`/`.kv` over a lazy source: emit the
-                    // index-based transform using the source position as the key,
-                    // bypassing the map/grep callback entirely.
-                    let idx = Value::Int(source_idx as i64);
-                    let produced: Vec<Value> = match index_transform.unwrap() {
-                        crate::value::IndexTransform::Pairs => {
-                            vec![Value::ValuePair(Box::new(idx), Box::new(elem))]
-                        }
-                        crate::value::IndexTransform::AntiPairs => {
-                            vec![Value::ValuePair(Box::new(elem), Box::new(idx))]
-                        }
-                        crate::value::IndexTransform::Kv => vec![idx, elem],
-                    };
-                    {
-                        let mut spec = list.lazy_pipe.as_ref().unwrap().lock().unwrap();
-                        spec.source_idx = source_idx + 1;
-                    }
-                    let mut cache = list.cache.lock().unwrap();
-                    cache.get_or_insert_with(Vec::new).extend(produced);
-                }
-                Some(elem) => {
-                    // Apply the stage with Interpreter-native dispatch so the callback
-                    // runs in *this* Interpreter (keeping locals/env coherent and letting
-                    // side effects reach the enclosing scope). A `grep` keeps the
-                    // element when the matcher is truthy; a `map` transforms it (a
-                    // `Slip` result contributes multiple elements).
-                    //
-                    // Loop control inside the callback is honored: `last` ends the
-                    // sequence (excluding the current element), `next` skips the
-                    // current element. This makes `(^Inf).grep({last if …})`
-                    // terminate instead of running forever.
-                    let applied: Result<Vec<Value>, RuntimeError> = if is_grep {
-                        match self.vm_grep_item_matches(&func, &elem) {
-                            Ok(true) => Ok(vec![elem]),
-                            Ok(false) => Ok(Vec::new()),
-                            Err(e) => Err(e),
-                        }
-                    } else {
-                        self.vm_call_on_value(func, vec![elem], None)
-                            .map(|v| match v {
-                                Value::Slip(items) => items.as_ref().clone(),
-                                v => vec![v],
-                            })
-                    };
-                    let produced: Vec<Value> = match applied {
-                        Ok(p) => p,
-                        Err(e) if e.is_last() => {
-                            // `last`: terminate the sequence, current element excluded.
-                            let mut spec = list.lazy_pipe.as_ref().unwrap().lock().unwrap();
-                            spec.done = true;
-                            drop(spec);
-                            let cache = list.cache.lock().unwrap();
-                            let c = cache.as_ref().cloned().unwrap_or_default();
-                            let n = needed.min(c.len());
-                            return Ok(c[..n].to_vec());
-                        }
-                        // `next`: skip the current element and continue.
-                        Err(e) if e.is_next() => Vec::new(),
-                        Err(e) => return Err(e),
-                    };
-                    {
-                        let mut spec = list.lazy_pipe.as_ref().unwrap().lock().unwrap();
-                        spec.source_idx = source_idx + 1;
-                    }
-                    let mut cache = list.cache.lock().unwrap();
-                    cache.get_or_insert_with(Vec::new).extend(produced);
-                }
-            }
-        }
-    }
-
-    /// Pull the `idx`-th element of a pipeline source, or `None` when the source
-    /// has fewer than `idx + 1` elements (finite source exhausted). Infinite
-    /// integer ranges always produce. Nested lazy pipelines / gathers are pulled
-    /// incrementally via [`Self::force_lazy_list_vm_n`].
-    fn pull_source_element(
-        &mut self,
-        source: &Value,
-        idx: usize,
-    ) -> Result<Option<Value>, RuntimeError> {
-        match source {
-            Value::Range(a, b)
-            | Value::RangeExcl(a, b)
-            | Value::RangeExclStart(a, b)
-            | Value::RangeExclBoth(a, b) => {
-                let start = match source {
-                    Value::RangeExclStart(..) | Value::RangeExclBoth(..) => a.saturating_add(1),
-                    _ => *a,
-                };
-                let inclusive = matches!(source, Value::Range(..) | Value::RangeExclStart(..));
-                let cur = match start.checked_add(idx as i64) {
-                    Some(v) => v,
-                    None => return Ok(None),
-                };
-                let in_bounds = if inclusive { cur <= *b } else { cur < *b };
-                if in_bounds {
-                    Ok(Some(Value::Int(cur)))
-                } else {
-                    Ok(None)
-                }
-            }
-            // Integer-start GenericRange (e.g. `^Inf`, `0..^Inf`): pull the
-            // idx-th element directly so an infinite range stays lazy instead of
-            // being materialized.
-            Value::GenericRange {
-                start,
-                end,
-                excl_start,
-                excl_end,
-            } if matches!(start.as_ref(), Value::Int(_)) => {
-                let s = match start.as_ref() {
-                    Value::Int(i) => *i,
-                    _ => unreachable!(),
-                };
-                let base = if *excl_start { s.saturating_add(1) } else { s };
-                let cur = match base.checked_add(idx as i64) {
-                    Some(v) => v,
-                    None => return Ok(None),
-                };
-                let end_f = end.to_f64();
-                let in_bounds = if end_f.is_infinite() && end_f.is_sign_positive() {
-                    true
-                } else {
-                    let cur_f = cur as f64;
-                    if *excl_end {
-                        cur_f < end_f
-                    } else {
-                        cur_f <= end_f
-                    }
-                };
-                Ok(in_bounds.then_some(Value::Int(cur)))
-            }
-            Value::Seq(items) | Value::Slip(items) => Ok(items.get(idx).cloned()),
-            Value::Array(items, _) => Ok(items.get(idx).cloned()),
-            Value::LazyList(ll) => {
-                let items = self.force_lazy_list_vm_n(ll, idx + 1)?;
-                Ok(items.get(idx).cloned())
-            }
-            // Other sources (non-integer GenericRange, etc.) are not gated into
-            // the lazy pipeline; materialize once and index.
-            other => {
-                let items = crate::runtime::value_to_list(other);
-                Ok(items.get(idx).cloned())
-            }
-        }
-    }
-
-    /// Force a LazyList into a Seq by evaluating the gather body.
-    /// Force a scan-based LazyList, computing up to `needed` elements.
-    /// Elements are computed incrementally and cached in the LazyList.
-    pub(super) fn force_scan_lazy_list(
-        &mut self,
-        list: &LazyList,
-        needed: usize,
-    ) -> Result<Vec<Value>, RuntimeError> {
-        let scan_mutex = match &list.scan_spec {
-            Some(s) => s,
-            None => return Ok(Vec::new()),
-        };
-
-        // Read current state under lock, then release before calling reduction methods
-        let (base_op, negate, source, mut acc, already, cached_len) = {
-            let spec = scan_mutex.lock().unwrap();
-            let cache_guard = list.cache.lock().unwrap();
-            let cached_len = cache_guard.as_ref().map_or(0, |v| v.len());
-            if cached_len >= needed {
-                return Ok(cache_guard.as_ref().unwrap()[..needed].to_vec());
-            }
-            (
-                spec.op.clone(),
-                spec.negate,
-                spec.source.clone(),
-                spec.accumulator.clone(),
-                spec.computed_count,
-                cached_len,
-            )
-        };
-
-        let callable = self.reduction_callable_for_op(&base_op);
-        let remaining = needed - cached_len;
-
-        // Collect new source values to iterate over
-        let new_values: Vec<Value> = match &source {
-            Value::Range(a, b) => {
-                let start = *a + already as i64;
-                let end = if *b == i64::MAX {
-                    *a + needed as i64
-                } else {
-                    *b
-                };
-                (start..=end).take(remaining).map(Value::Int).collect()
-            }
-            Value::RangeExcl(a, b) => {
-                let start = *a + already as i64;
-                let end = if *b == i64::MAX {
-                    *a + needed as i64
-                } else {
-                    *b
-                };
-                (start..end).take(remaining).map(Value::Int).collect()
-            }
-            Value::RangeExclStart(a, b) => {
-                let first = *a + 1;
-                let start = first + already as i64;
-                let end = if *b == i64::MAX {
-                    first + needed as i64
-                } else {
-                    *b
-                };
-                (start..=end).take(remaining).map(Value::Int).collect()
-            }
-            Value::RangeExclBoth(a, b) => {
-                let first = *a + 1;
-                let start = first + already as i64;
-                let end = if *b == i64::MAX {
-                    first + needed as i64
-                } else {
-                    *b
-                };
-                (start..end).take(remaining).map(Value::Int).collect()
-            }
-            Value::GenericRange {
-                start,
-                end,
-                excl_start,
-                ..
-            } => {
-                let end_f = end.to_f64();
-                let is_infinite = end_f.is_infinite() && end_f.is_sign_positive();
-                let start_i = start.as_ref().to_f64() as i64;
-                let first_i = if *excl_start { start_i + 1 } else { start_i };
-                let iter_start = first_i + already as i64;
-                let iter_end = if is_infinite {
-                    iter_start + remaining as i64
-                } else {
-                    (end_f as i64).min(iter_start + remaining as i64)
-                };
-                (iter_start..=iter_end)
-                    .take(remaining)
-                    .map(Value::Int)
-                    .collect()
-            }
-            _ => {
-                let items = crate::runtime::utils::value_to_list(&source);
-                items.into_iter().skip(already).take(remaining).collect()
-            }
-        };
-
-        // Compute new scan elements (no locks held)
-        let mut new_out: Vec<Value> = Vec::new();
-        let mut computed = already;
-
-        for val in new_values {
-            acc = Some(match acc.take() {
-                None => {
-                    new_out.push(val.clone());
-                    val
-                }
-                Some(prev) => {
-                    let call_args = vec![prev, val];
-                    let v =
-                        self.reduction_step_with_args(&base_op, callable.as_ref(), call_args)?;
-                    let v = if negate { Value::Bool(!v.truthy()) } else { v };
-                    new_out.push(v.clone());
-                    v
-                }
-            });
-            computed += 1;
-        }
-
-        // Update spec and cache under lock
-        {
-            let mut spec = scan_mutex.lock().unwrap();
-            spec.accumulator = acc;
-            spec.computed_count = computed;
-
-            let mut cache_guard = list.cache.lock().unwrap();
-            let out = cache_guard.get_or_insert_with(Vec::new);
-            out.extend(new_out);
-
-            if out.len() >= needed {
-                Ok(out[..needed].to_vec())
-            } else {
-                Ok(out.clone())
-            }
-        }
-    }
-
-    fn force_lazy_if_needed(&mut self, val: Value) -> Result<Value, RuntimeError> {
-        if let Value::LazyList(ll) = &val {
-            let items = self.force_lazy_list_vm(ll)?;
-            Ok(Value::Seq(std::sync::Arc::new(items)))
-        } else {
-            Ok(val)
-        }
-    }
-
-    pub(super) fn eval_binary_with_junctions(
-        &mut self,
-        left: Value,
-        right: Value,
-        f: fn(&mut Interpreter, Value, Value) -> Result<Value, RuntimeError>,
-    ) -> Result<Value, RuntimeError> {
-        // Auto-FETCH Proxy containers in binary operations
-        let left = loan_env!(self, auto_fetch_proxy(&left))?;
-        let right = loan_env!(self, auto_fetch_proxy(&right))?;
-        // Decontainerize Scalar wrappers
-        let left = left.descalarize().clone();
-        let right = right.descalarize().clone();
-        if let (
-            Value::Junction {
-                kind: left_kind,
-                values: _,
-            },
-            Value::Junction {
-                kind: right_kind,
-                values: right_values,
-            },
-        ) = (&left, &right)
-            && Self::thread_right_first(left_kind, right_kind)
-        {
-            let results: Result<Vec<Value>, RuntimeError> = right_values
-                .iter()
-                .cloned()
-                .map(|v| self.eval_binary_with_junctions(left.clone(), v, f))
-                .collect();
-            return Ok(Value::junction(right_kind.clone(), results?));
-        }
-        if let Value::Junction { kind, values } = left {
-            let results: Result<Vec<Value>, RuntimeError> = values
-                .iter()
-                .cloned()
-                .map(|v| self.eval_binary_with_junctions(v, right.clone(), f))
-                .collect();
-            return Ok(Value::junction(kind, results?));
-        }
-        if let Value::Junction { kind, values } = right {
-            let results: Result<Vec<Value>, RuntimeError> = values
-                .iter()
-                .cloned()
-                .map(|v| self.eval_binary_with_junctions(left.clone(), v, f))
-                .collect();
-            return Ok(Value::junction(kind, results?));
-        }
-        // Force LazyList values before arithmetic/comparison operations
-        let left = self.force_lazy_if_needed(left)?;
-        let right = self.force_lazy_if_needed(right)?;
-        f(self, left, right)
-    }
-
-    /// Smartmatch with junction threading but WITHOUT forcing lazy values.
-    /// For `!~~` (negate=true), we compute `~~` first and then negate the
-    /// collapsed result.  Raku defines `$x !~~ $y` as `not ($x ~~ $y)`,
-    /// where `not` collapses junctions before negating.
-    #[allow(dead_code)]
-    pub(super) fn eval_smartmatch_with_junctions(
-        &mut self,
-        left: Value,
-        right: Value,
-        negate: bool,
-    ) -> Result<Value, RuntimeError> {
-        self.eval_smartmatch_with_junctions_ex(left, right, negate, false)
-    }
-
-    /// Extended smartmatch with junction threading.
-    /// `rhs_is_match_regex` indicates the RHS was originally `m//`, which
-    /// changes the failure return from Nil to False.
-    pub(super) fn eval_smartmatch_with_junctions_ex(
-        &mut self,
-        left: Value,
-        right: Value,
-        negate: bool,
-        rhs_is_match_regex: bool,
-    ) -> Result<Value, RuntimeError> {
-        // For !~~, compute ~~ first, then negate the collapsed boolean.
-        if negate {
-            let match_result =
-                self.eval_smartmatch_with_junctions_ex(left, right, false, rhs_is_match_regex)?;
-            let bool_val = match_result.truthy();
-            return Ok(Value::Bool(!bool_val));
-        }
-        // When RHS is the Junction type object, don't auto-thread LHS.
-        // $junction ~~ Junction should return True (a Junction isa Junction).
-        // Also applies to Mu (the supertype of Junction).
-        if matches!(&right, Value::Package(name) if matches!(name.resolve().as_str(), "Junction" | "Mu"))
-            && matches!(&left, Value::Junction { .. })
-        {
-            return self.smart_match_op(left, right, rhs_is_match_regex);
-        }
-        // Helper: check if a value is a regex (for junction collapse decisions)
-        let is_regex_value = |v: &Value| {
-            matches!(
-                v,
-                Value::Regex(_)
-                    | Value::RegexWithAdverbs { .. }
-                    | Value::Routine { is_regex: true, .. }
-            )
-        };
-        if let (
-            Value::Junction {
-                kind: left_kind,
-                values: _,
-            },
-            Value::Junction {
-                kind: right_kind,
-                values: right_values,
-            },
-        ) = (&left, &right)
-            && Self::thread_right_first(left_kind, right_kind)
-        {
-            let results: Result<Vec<Value>, RuntimeError> = right_values
-                .iter()
-                .cloned()
-                .map(|v| {
-                    self.eval_smartmatch_with_junctions_ex(
-                        left.clone(),
-                        v,
-                        false,
-                        rhs_is_match_regex,
-                    )
-                })
-                .collect();
-            // Smartmatch collapses junctions to Bool
-            let junction = Value::junction(right_kind.clone(), results?);
-            return Ok(Value::Bool(junction.truthy()));
-        }
-        if let Value::Junction { kind, values } = left {
-            // When RHS is a non-junction regex and LHS is a junction,
-            // return the Junction of Match/Nil results without collapsing.
-            // For all other cases, collapse to Bool.
-            let keep_junction = is_regex_value(&right);
-            let results: Result<Vec<Value>, RuntimeError> = values
-                .iter()
-                .cloned()
-                .map(|v| {
-                    self.eval_smartmatch_with_junctions_ex(
-                        v,
-                        right.clone(),
-                        false,
-                        rhs_is_match_regex,
-                    )
-                })
-                .collect();
-            let junction = Value::junction(kind, results?);
-            if keep_junction {
-                return Ok(junction);
-            }
-            return Ok(Value::Bool(junction.truthy()));
-        }
-        if let Value::Junction { kind, values } = right {
-            // Evaluate junction elements with short-circuit semantics:
-            // All: if any element is False, stop early (don't evaluate remaining).
-            // Any: if any element is True, stop early.
-            // One: always evaluate all (no short-circuit possible).
-            let mut results = Vec::with_capacity(values.len());
-            for v in values.iter().cloned() {
-                let r = self.eval_smartmatch_with_junctions_ex(
-                    left.clone(),
-                    v,
-                    false,
-                    rhs_is_match_regex,
-                )?;
-                let is_truthy = r.truthy();
-                results.push(r);
-                match &kind {
-                    crate::value::JunctionKind::All if !is_truthy => break, // short-circuit: All fails fast
-                    crate::value::JunctionKind::Any if is_truthy => break, // short-circuit: Any succeeds fast
-                    _ => {}
-                }
-            }
-            // Smartmatch collapses junctions to Bool
-            let junction = Value::junction(kind, results);
-            return Ok(Value::Bool(junction.truthy()));
-        }
-        self.smart_match_op(left, right, rhs_is_match_regex)
-    }
-
-    pub(super) fn smart_match_op(
-        &mut self,
-        left: Value,
-        right: Value,
-        rhs_is_match_regex: bool,
-    ) -> Result<Value, RuntimeError> {
-        // When RHS is Whatever, autoprime: return a WhateverCode that takes
-        // one argument and smartmatches LHS against it.
-        // In Raku, `$x ~~ *` produces `-> $a { $x ~~ $a }`.
-        if matches!(&right, Value::Whatever) {
-            use crate::ast::{Expr, Stmt};
-            use crate::env::Env;
-            let mut env = Env::new();
-            env.insert(
-                "__mutsu_callable_type".to_string(),
-                Value::str_from("WhateverCode"),
-            );
-            // Capture the LHS value in the closure environment
-            env.insert("__wc_sm_lhs".to_string(), left);
-            let param = "__wc_0".to_string();
-            let body = vec![Stmt::Expr(Expr::Binary {
-                left: Box::new(Expr::Var("__wc_sm_lhs".to_string())),
-                op: crate::token_kind::TokenKind::SmartMatch,
-                right: Box::new(Expr::Var(param.clone())),
-            })];
-            return Ok(Value::make_sub(
-                Symbol::intern("GLOBAL"),
-                Symbol::intern("<whatevercode-smartmatch>"),
-                vec![param],
-                Vec::new(),
-                body,
-                false,
-                env,
-            ));
-        }
-        let is_regex = matches!(
-            &right,
-            Value::Regex(_)
-                | Value::RegexWithAdverbs { .. }
-                | Value::Routine { is_regex: true, .. }
-        );
-        let matched = self.vm_smart_match(&left, &right);
-        // Check for pending regex security error (set by regex parse/match)
-        if let Some(err) = crate::runtime::Interpreter::take_pending_regex_error() {
-            return Err(err);
-        }
-        // Check for pending dispatch error (e.g., from Any ~~ Pair method call)
-        if let Some(err) = self.take_pending_dispatch_error() {
-            return Err(err);
-        }
-        if is_regex {
-            // When $/ is a Junction (from :nth with junction argument),
-            // the ~~ operator collapses the result to a Bool.
-            let slash = self.env().get("/").cloned().unwrap_or(Value::Nil);
-            if matches!(&slash, Value::Junction { .. }) {
-                Ok(Value::Bool(matched))
-            } else if matched {
-                // For regex smartmatch, return the Match object (from $/) or Nil
-                Ok(slash)
-            } else if rhs_is_match_regex && matches!(&slash, Value::Nil) {
-                // Failed m// (non-global) returns False, not Nil.
-                // But m:g// returns an empty list from $/, so we check that
-                // $/ is Nil before returning False.
-                Ok(Value::Bool(false))
-            } else {
-                // Failed bare // returns Nil; m:g// returns $/ (empty list)
-                Ok(slash)
-            }
-        } else {
-            Ok(Value::Bool(matched))
-        }
-    }
-
-    /// Decont `:=`-bound `ContainerRef` cells inside an Array value.
-    ///
-    /// When an array element has been bound (`@a[1] := $var`), the element
-    /// holds a shared `ContainerRef` cell. This method replaces every such
-    /// cell with its current value, so that callers that snapshot or iterate
-    /// elements (assignment copies, stringification, `say`, `gist`) see the
-    /// live value instead of the cell.
-    pub(super) fn resolve_bound_array_elements(&self, val: Value) -> Value {
-        if let Value::Array(ref items, kind) = val {
-            let needs_resolve = items.iter().any(Value::is_container_ref);
-            if !needs_resolve {
-                return val;
-            }
-            let resolved: Vec<Value> = items
-                .iter()
-                .map(|v| match v {
-                    Value::ContainerRef(cell) => cell.lock().unwrap().clone(),
-                    other => other.clone(),
-                })
-                .collect();
-            Value::Array(
-                std::sync::Arc::new(crate::value::ArrayData::new(resolved)),
-                kind,
-            )
-        } else {
-            val
-        }
-    }
-
-    /// Extend an infinite closure-based sequence to at least `needed` elements
-    /// by re-invoking its generator closure over the growing element history.
-    /// Returns whatever is available (possibly fewer than `needed`) once the
-    /// generator signals termination.
-    fn extend_closure_sequence(
-        &mut self,
-        list: &LazyList,
-        needed: usize,
-    ) -> Result<Vec<Value>, RuntimeError> {
-        // Fast path: already cached enough.
-        {
-            let cache = list.cache.lock().unwrap();
-            if let Some(cached) = cache.as_ref()
-                && cached.len() >= needed
-            {
-                return Ok(cached[..needed].to_vec());
-            }
-        }
-
-        // Snapshot the current history without holding the cache lock across
-        // user-code execution.
-        let mut history = {
-            let cache = list.cache.lock().unwrap();
-            cache.as_ref().cloned().unwrap_or_default()
-        };
-
-        let state_mutex = list.closure_seq.as_ref().unwrap();
-        let mut guard = state_mutex.lock().unwrap();
-        let state = &mut *guard;
-        let generator = state.generator.clone();
-
-        while history.len() < needed && !state.finished {
-            let precompiled = state
-                .precompiled
-                .as_ref()
-                .map(|(c, f)| (c.as_ref(), f.as_ref()));
-            match self.sequence_closure_step(
-                &generator,
-                &history,
-                precompiled,
-                &mut state.closure_env,
-                false,
-            )? {
-                Some(v) => history.push(v),
-                None => {
-                    state.finished = true;
-                    break;
-                }
-            }
-        }
-
-        // Publish the extended history back to the cache.
-        *list.cache.lock().unwrap() = Some(history.clone());
-        let take = needed.min(history.len());
-        Ok(history[..take].to_vec())
-    }
-
-    /// Extend a sequence-spec lazy list's cache to at least `needed` elements.
-    /// This generates new elements using the sequence spec (arithmetic/geometric)
-    /// without needing any Interpreter or interpreter context.
-    fn extend_sequence_cache(
-        list: &LazyList,
-        spec: &crate::value::SequenceSpec,
-        needed: usize,
-    ) -> Result<Vec<Value>, RuntimeError> {
-        let mut cache = list.cache.lock().unwrap();
-        let items = cache.get_or_insert_with(Vec::new);
-        if items.len() >= needed {
-            return Ok(items[..needed].to_vec());
-        }
-        // Generate more elements
-        while items.len() < needed {
-            let last = items.last().cloned().unwrap_or(Value::Int(0));
-            let next = match spec {
-                crate::value::SequenceSpec::Arithmetic { step, all_int } => {
-                    if *all_int {
-                        if let Value::Int(n) = last {
-                            Value::Int(n + step)
-                        } else {
-                            let n = last.to_f64();
-                            Value::Num(n + *step as f64)
-                        }
-                    } else {
-                        let n = last.to_f64();
-                        Value::Num(n + *step as f64)
-                    }
-                }
-                crate::value::SequenceSpec::GeometricRat { num, den } => {
-                    if let Value::Int(n) = last {
-                        let result = n * num;
-                        if result % den == 0 {
-                            Value::Int(result / den)
-                        } else {
-                            Value::Rat(result * num, *den)
-                        }
-                    } else {
-                        let n = last.to_f64();
-                        Value::Num(n * (*num as f64) / (*den as f64))
-                    }
-                }
-                crate::value::SequenceSpec::Geometric { ratio } => {
-                    let n = last.to_f64();
-                    Value::Num(n * ratio)
-                }
-            };
-            items.push(next);
-        }
-        Ok(items[..needed].to_vec())
-    }
 }
